requirements.txt
Flask==2.3.3
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.0.1
SQLAlchemy==2.0.23
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
ultralytics==8.0.196
requests==2.31.0
scipy==1.11.4